@@ -2,14 +2,14 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
 class SizeConfig {
-  static double _screenWidth;
-  static double _screenHeight;
+  static double? _screenWidth;
+  static double? _screenHeight;
   static double _blockWidth = 0;
   static double _blockHeight = 0;
-  static double textMultiplier;
-  static double imageSizeMultiplier;
-  static double heightMultiplier;
-  static double widthMultiplier;
+  static double textMultiplier = 0;
+  static double imageSizeMultiplier = 0;
+  static double heightMultiplier = 0;
+  static double widthMultiplier = 0;
   static bool isPortrait = true;
   static bool isMobilePortrait = false;
 
@@ -18,7 +18,7 @@ class SizeConfig {
       _screenWidth = constraints.maxWidth;
       _screenHeight = constraints.maxHeight;
       isPortrait = true;
-      if (_screenWidth < 450) {
+      if (_screenWidth! < 450) {
         isMobilePortrait = true;
       }
     } else {
@@ -27,8 +27,8 @@ class SizeConfig {
       isPortrait = false;
       isMobilePortrait = false;
     }
-    _blockWidth = _screenWidth / 100;
-    _blockHeight = _screenHeight / 100;
+    _blockWidth = _screenWidth! / 100;
+    _blockHeight = _screenHeight! / 100;
     textMultiplier = _blockHeight;
     imageSizeMultiplier = _blockWidth;
     heightMultiplier = _blockHeight;
