@@ -5,8 +5,7 @@ class RandomNumberGenerator {
   static const String _chars =
       '0123456789AaBbCcDdEeFfGgHhIiJjKkMmNnOoPpQqRrSsTtUuVvWwXxYyZz';
 
-  String randomStringGenerator(int length) {
-    return String.fromCharCodes(
+  String randomStringGenerator(int length) => String.fromCharCodes(
       Iterable<int>.generate(
         length,
         (_) => _chars.codeUnitAt(
@@ -14,5 +13,4 @@ class RandomNumberGenerator {
         ),
       ),
     );
-  }
 }
